Changed the flowbroker to guarantee the events are processed in order
Solution: Messages are moved from the Kafka 'device-data' subject
to the queue 'event_queue' in Rabbitmq. Then they are pre-processed
one at a time and tasks are pushed to queue 'task_queue' to be processed
also one at a time. Once a task is processed if new tasks are spawned,
they have higher priorities.